code/pyproject.toml
[tool.poetry]
name = "api-product"
version = "0.1.0"
description = "api that allows you to manage products"
authors = ["smc"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.13.2"
pytest-mock = "^3.14.1"
asyncio = "^3.4.3"
uuid = "^1.30"
fastapi = "^0.116.0"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"
uvicorn = "^0.35.0"
jose = "^1.0.0"
bcrypt = "4.0.1"
structlog = "^25.4.0"
fastapi-pagination = "^0.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
